+++ R/team_members.R
@@ -19,19 +19,18 @@ team_members = function(org, team = org_teams(org), team_type = c("name", "slug"
   arg_is_chr(team)
   team_type = match.arg(team_type)
 
-  if (team_type == "name")
-    team = team_slug_lookup(org, team)
+  slug = if (team_type == "name") team_slug_lookup(org, team) else team
 
-  check_team_slug(team)
+  check_team_slug(slug)
 
-  purrr::map_dfr(
-    team,
-    function(team) {
+  purrr::map2_dfr(
+    team, slug,
+    function(team, slug) {
 
-      if (is.na(team)) {
+      if (is.na(slug)) {
         res = NULL
       } else {
-        res = purrr::safely(github_api_team_members)(org, team)
+        res = purrr::safely(github_api_team_members)(org, slug)
 
         status_msg(
           res,
@@ -39,17 +38,16 @@ team_members = function(org, team = org_teams(org), team_type = c("name", "slug"
         )
       }
 
-      if (failed(res) | empty_result(res)) {
-        tibble::tibble(
-          team = character(),
-          user = character()
-        )
-      } else {
-        tibble::tibble(
-          team = team,
-          user = purrr::map_chr(result(res), "login"),
-        )
-      }
+      members = if (failed(res) | empty_result(res))
+        character()
+      else
+        purrr::map_chr(result(res), "login")
+
+      tibble::tibble(
+        team = team,
+        slug = slug,
+        user = members
+      )
     }
   )
 }