+++ R/team_repos.R
@@ -15,30 +15,28 @@ team_repos = function(org, team = org_teams(org), team_type = c("name", "slug"))
   arg_is_chr(team)
   team_type = match.arg(team_type)
 
-  if (team_type == "name")
-    team = team_slug_lookup(org, team)
+  slug = if (team_type == "name") team_slug_lookup(org, team) else team
 
-  check_team_slug(team)
+  check_team_slug(slug)
 
-  purrr::map_dfr(
-    team,
-    function(team) {
-      if (is.na(team))
+  purrr::map2_dfr(
+    team, slug,
+    function(team, slug) {
+      if (is.na(slug))
         res = NULL
       else
-        res = purrr::safely(github_api_team_repos)(org, team)
+        res = purrr::safely(github_api_team_repos)(org, slug)
 
-      if (failed(res) | empty_result(res)) {
-        tibble::tibble(
-          team = character(),
-          repo = character()
-        )
-      } else {
-        tibble::tibble(
-          team = team,
-          repo = purrr::map_chr(result(res), "full_name")
-        )
-      }
+      repos = if (failed(res) | empty_result(res))
+        character()
+      else
+        purrr::map_chr(result(res), "full_name")
+
+      tibble::tibble(
+        team = team,
+        slug = slug,
+        repo = repos
+      )
     }
   )
 }